renames util directory to utils
